S4 class containing recipes, their ingredients and meal plans

A book of recipes from which meal plans and shopping lists can be generated.

RecipeBook(names, ingredients)

# S4 method for RecipeBook
show(object)

# S4 method for RecipeBook
names(x)

# S4 method for RecipeBook
length(x)

recipes(object)

meal_plan(object)

meal_plan(object) <- value

favourites(x)

favourites(object, ...) <- value

# S4 method for RecipeBook,numeric
favourites(object) <- value

# S4 method for RecipeBook,logical
favourites(object) <- value

# S4 method for RecipeBook,character
favourites(object, con = stdin()) <- value

Arguments

names
character() containing the name of of recipes.

ingredients
character() or list() of Ingredient-class objects. If a character(), will be read in using read_ingredients().

object
a RecipeBook-class object.

x
a RecipeBook-class object.

value
either a integer() specifying the index of the rows to favourite, an logical() to set as the "fav" column, or the character() "manual" to specify setting the favourites manually.

...
additional arguments passed on the favourite<- methods.

con
a connection object or a character string.

Slots

recipes

tibble::tibble() storing the names, ingredients and favourites status of recipes.

meal_plan

tibble::tibble() storing a meal plan, created by recipes using create_meal_plan().

Constructor

RecipeBook(names, ingredients) creates an object of RecipeBook-class.

Displaying

show(recipebook) prints the recipes as a tibble::tibble().

Getters

names(recipebook) obtains the names of the recipes.

length(recipebook) obtains the number of recipes.

recipes(recipebook) obtains the recipes as a tibble::tibble().

meal_plan(recipebook) obtains the meal_plan as a tibble::tibble().

favourites(recipebook) obtains the recipes marked as favourites.

Setters

meal_plan(recipebook) <- meal_plan stores the meal plan inside a RecipeBook-class object.

favourites(recipebook) <- 1 set favourite recipes by index.

favourites(recipebook) <- TRUE set favourite recipes by logical.

favourites(recipebook) <- "manual" set favourite recipes manually.