Create a meal plan

Create a meal plan

create_meal_plan(
  recipebook,
  days = c("Mon", "Tues", "Wed", "Thurs", "Fri", "Sat", "Sun"),
  meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE
)

Arguments

recipebook
a RecipeBook-class object.

days
character() which days of the week to plan for.

meals
character() which meals to plan for, out of "Lunch" and "Dinner".

method
character() the method to use when creating a meal plan.

fav_only
logical() whether to create a meal plan only using favourite recipes.

Value

a RecipeBook-class object containing the created meal_plan.

Examples

meal_plan(RecipeBook_example) <- create_meal_plan(RecipeBook_example)

meal_plan(RecipeBook_example)

#> # A tibble: 14 × 3
#>    day   meal   recipe_index
#>    <fct> <fct>         <int>
#>  1 Mon   Lunch            13
#>  2 Mon   Dinner           23
#>  3 Tues  Lunch            12
#>  4 Tues  Dinner            5
#>  5 Wed   Lunch             6
#>  6 Wed   Dinner           15
#>  7 Thurs Lunch            21
#>  8 Thurs Dinner            4
#>  9 Fri   Lunch            22
#> 10 Fri   Dinner            9
#> 11 Sat   Lunch            25
#> 12 Sat   Dinner           16
#> 13 Sun   Lunch             8
#> 14 Sun   Dinner           26