An example recipe book

A data set that has been created using autorecipes::create_recipebook() containing 26 example recipes. The recipes originate from recipe boxes that I've ordered from Riverford. Riverford are a wonderful UK company producing sustainable, delicious, locally-sourced food - check them out!

recipebook_example

Format

A tibble::tibble() with 26 rows and 2 columns:

names

character() recipe name.

ingredients

character() name, quantity and type of ingredient, stored as a list() of Ingredients-class objects.

Source

generated using data-raw/recipebook_example.R