Create a meal plan

Create a meal plan

create_meal_plan(
  recipebook,
  which_days = weekdays(),
  which_meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE,
  set_last_eaten = TRUE
)

Arguments

recipebook
a RecipeBook-class object.

which_days
character() which days of the week to plan for.

which_meals
character() which meals to plan for, out of "Lunch" and "Dinner".

method
character() the method to use when creating a meal plan.

fav_only
logical() whether to create a meal plan only using favourite recipes.

set_last_eaten
logical() whether to save the recipes were eaten today. Used to optimise the meal planning algorithm when the method is set to "auto".

Value

a RecipeBook-class object containing the created meal_plan.

Examples

recipebook <- create_meal_plan(recipebook_example)

meal_plan(recipebook)
#> # A tibble: 14 × 7
#>    day   meal   recipe_index names                  ingredients fav   last_eaten
#>    <fct> <fct>         <int> <chr>                  <list>      <lgl> <date>    
#>  1 Mon   Lunch            22 New Potato Tagine      <Ingrdnts>  FALSE 2021-10-14
#>  2 Mon   Dinner           14 Roast Mushroom, Creme… <Ingrdnts>  FALSE 2021-10-14
#>  3 Tues  Lunch             4 Chickpea & Sweet Pota… <Ingrdnts>  FALSE 2021-10-14
#>  4 Tues  Dinner           13 Creamy Chicken & Much… <Ingrdnts>  FALSE 2021-10-14
#>  5 Wed   Lunch            11 Pink Grapefruit, Mint… <Ingrdnts>  FALSE 2021-10-14
#>  6 Wed   Dinner           17 Aubergine Schnitzel w… <Ingrdnts>  FALSE 2021-10-14
#>  7 Thurs Lunch             2 Crispy Duck with Fava… <Ingrdnts>  FALSE 2021-10-14
#>  8 Thurs Dinner            3 Chicken with Rosemary… <Ingrdnts>  FALSE 2021-10-14
#>  9 Fri   Lunch             9 Squash, Corn & Tomati… <Ingrdnts>  FALSE 2021-10-14
#> 10 Fri   Dinner           25 Korean Beef & Kimchi … <Ingrdnts>  FALSE 2021-10-14
#> 11 Sat   Lunch             6 Chipotle Bean Taco Sa… <Ingrdnts>  FALSE 2021-10-14
#> 12 Sat   Dinner           19 Beef & Spinach Lasagne <Ingrdnts>  FALSE 2021-10-14
#> 13 Sun   Lunch            18 Normandy Pork & Cider… <Ingrdnts>  FALSE 2021-10-14
#> 14 Sun   Dinner           10 Chipotle, Corn & Chee… <Ingrdnts>  FALSE 2021-10-14