Create a customizable meal plan

create_meal_plan() expects a RecipeBook-class object as input. The available options for customizing the meal plan include; the days of the week and the meals ("Lunch" or "Dinner") to plan for, the method to use and whether to only use favourite recipes.

create_meal_plan(
  recipebook,
  which_days = weekdays(),
  which_meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE,
  set_last_eaten = TRUE
)

Arguments

recipebook
a RecipeBook-class object.

which_days
character() which days of the week to plan for, in the format "Mon"

which_meals
character() which meals to plan for, one of "Lunch" or "Dinner".

method
character() the method to use when creating a meal plan, one of "auto", or "random".

fav_only
logical() whether to only pick from favourite recipes.

set_last_eaten
logical() whether to record the that selected recipes were used for a meal plan today. This info is taken into account when optimising the meal planning algorithm if method is set to "auto".

Value

a RecipeBook-class object containing the created meal_plan.

Examples

recipebook <- create_meal_plan(recipebook_example)

meal_plan(recipebook)
#> # A tibble: 14 × 7
#>    day   meal   recipe_index names                  ingredients fav   last_eaten
#>    <fct> <fct>         <int> <chr>                  <list>      <lgl> <date>    
#>  1 Mon   Lunch             7 Steak Tagliata & Roas… <Ingrdnts>  FALSE 2021-10-20
#>  2 Mon   Dinner            5 Gaucho Steaks with Ch… <Ingrdnts>  FALSE 2021-10-20
#>  3 Tues  Lunch            13 Creamy Chicken & Much… <Ingrdnts>  FALSE 2021-10-20
#>  4 Tues  Dinner           26 Afghan Spliced Squash  <Ingrdnts>  FALSE 2021-10-20
#>  5 Wed   Lunch            16 Creamy Pollock, Samph… <Ingrdnts>  FALSE 2021-10-20
#>  6 Wed   Dinner            2 Crispy Duck with Fava… <Ingrdnts>  FALSE 2021-10-20
#>  7 Thurs Lunch             1 Akoori Eggs with Spin… <Ingrdnts>  FALSE 2021-10-20
#>  8 Thurs Dinner           14 Roast Mushroom, Creme… <Ingrdnts>  FALSE 2021-10-20
#>  9 Fri   Lunch            19 Beef & Spinach Lasagne <Ingrdnts>  FALSE 2021-10-20
#> 10 Fri   Dinner            9 Squash, Corn & Tomati… <Ingrdnts>  FALSE 2021-10-20
#> 11 Sat   Lunch            22 New Potato Tagine      <Ingrdnts>  FALSE 2021-10-20
#> 12 Sat   Dinner           12 Honey & Sesame Buffal… <Ingrdnts>  FALSE 2021-10-20
#> 13 Sun   Lunch             8 Sweetcorn & Spring On… <Ingrdnts>  FALSE 2021-10-20
#> 14 Sun   Dinner           21 Chickpea Falafels wit… <Ingrdnts>  FALSE 2021-10-20