Create a customizable meal plan

create_meal_plan() expects a RecipeBook-class object as input. The available options for customizing the meal plan include; the days of the week and the meals ("Lunch" or "Dinner") to plan for, the method to use and whether to only use favourite recipes.

create_meal_plan(
  recipebook,
  which_days = weekdays(),
  which_meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE,
  set_last_eaten = TRUE
)

Arguments

recipebook

a RecipeBook-class object.

which_days

character() which days of the week to plan for, in the format "Mon"

which_meals

character() which meals to plan for, one of "Lunch" or "Dinner".

method

character() the method to use when creating a meal plan, one of "auto", or "random".

fav_only

logical() whether to only pick from favourite recipes.

set_last_eaten

logical() whether to record the that selected recipes were used for a meal plan today. This info is taken into account when optimising the meal planning algorithm if method is set to "auto".

Value

a RecipeBook-class object containing the created meal_plan.

Examples

recipebook <- create_meal_plan(recipebook_example)

meal_plan(recipebook)
#> # A tibble: 14 × 7
#>    day   meal   recipe_index names                  ingredients fav   last_eaten
#>    <fct> <fct>         <int> <chr>                  <list>      <lgl> <date>    
#>  1 Mon   Lunch            11 Pink Grapefruit, Mint… <Ingrdnts>  FALSE 2021-12-30
#>  2 Mon   Dinner            9 Squash, Corn & Tomati… <Ingrdnts>  FALSE 2021-12-30
#>  3 Tues  Lunch             4 Chickpea & Sweet Pota… <Ingrdnts>  FALSE 2021-12-30
#>  4 Tues  Dinner           26 Afghan Spliced Squash  <Ingrdnts>  FALSE 2021-12-30
#>  5 Wed   Lunch             5 Gaucho Steaks with Ch… <Ingrdnts>  FALSE 2021-12-30
#>  6 Wed   Dinner           14 Roast Mushroom, Creme… <Ingrdnts>  FALSE 2021-12-30
#>  7 Thurs Lunch            12 Honey & Sesame Buffal… <Ingrdnts>  FALSE 2021-12-30
#>  8 Thurs Dinner           25 Korean Beef & Kimchi … <Ingrdnts>  FALSE 2021-12-30
#>  9 Fri   Lunch            10 Chipotle, Corn & Chee… <Ingrdnts>  FALSE 2021-12-30
#> 10 Fri   Dinner            8 Sweetcorn & Spring On… <Ingrdnts>  FALSE 2021-12-30
#> 11 Sat   Lunch            15 Globe Arichoke & Pine… <Ingrdnts>  FALSE 2021-12-30
#> 12 Sat   Dinner           17 Aubergine Schnitzel w… <Ingrdnts>  FALSE 2021-12-30
#> 13 Sun   Lunch            16 Creamy Pollock, Samph… <Ingrdnts>  FALSE 2021-12-30
#> 14 Sun   Dinner           13 Creamy Chicken & Much… <Ingrdnts>  FALSE 2021-12-30